Add files via upload
includes: 1: a small python programme that cleans data and also group categorises, it so it can be used in powerBI for slicer etc, 2 the actual csv cleaned dataset for the rest of the team to use to generate powerBI charts so we are consistent, 3   It also has another csv file which python generated for charts and graphs in the origninal file Jupyter notebook( note:  for improvement could have done one python programme that cleaned the data and group categorised it instead of two seperate programme but could not get it too work on VScode) 4 :Juypter notebook that generates graphs and charts using python, although not interative are useful. 5 myinteractive dahboardfor powerBI so we as a group can decide which ones to use after everyone submitted.